add preview page, cleanup components
