--- app/(personal)/components/ProjectPage.tsx
@@ -1,3 +1,0 @@
-export function ProjectPage() {
-  return <div>heu</div>
-}
