+++ app/(personal)/components/ThemeToggle.tsx
@@ -2,7 +2,7 @@
 import { useTheme } from 'next-themes'
 import { useEffect, useState } from 'react'
 
-export default function ToggleButton() {
+export function ThemeToggle() {
   const [mounted, setMounted] = useState(false)
   const { theme, setTheme } = useTheme()
