+++ app/(personal)/components/index.tsx
@@ -2,4 +2,4 @@ export { Footer } from './Footer'
 export { Header } from './Header'
 export { Navbar } from './Navbar'
 export { Project } from './Project'
-export { ProjectPage } from './ProjectPage'
+export { ProjectListItem } from './ProjectListItem'
